Adds a leaner OpenWhisk configuration. (#4216)
This commit introduces a leaner configuration of the deployment that eschews kafka & zookeeper. It merges the controller and invoker together and uses an in-memory bus to communicate between the controller and invoker. The leaner configuration reduces the amount of memory needed by a deploymen.
