+++ ansible/group_vars/all
@@ -2,6 +2,7 @@
 # license agreements; and to You under the Apache License, Version 2.0.
 
 mode: deploy
+lean: false
 prompt_user: true
 openwhisk_home: "{{ lookup('env', 'OPENWHISK_HOME') | default(playbook_dir ~ '/..', true) }}"
 openwhisk_cli_home: "{{ lookup('env', 'OPENWHISK_CLI') | default(openwhisk_home ~ '/../incubator-openwhisk-cli', true) }}"
