+++ core/controller/src/main/scala/org/apache/openwhisk/core/loadBalancer/ShardingContainerPoolBalancer.scala
@@ -19,34 +19,27 @@ package org.apache.openwhisk.core.loadBalancer
 
 import akka.actor.ActorRef
 import akka.actor.ActorRefFactory
-import java.nio.charset.StandardCharsets
 import java.util.concurrent.ThreadLocalRandom
-import java.util.concurrent.atomic.LongAdder
 
 import akka.actor.{Actor, ActorSystem, Cancellable, Props}
 import akka.cluster.ClusterEvent._
 import akka.cluster.{Cluster, Member, MemberStatus}
-import akka.event.Logging.InfoLevel
 import akka.management.AkkaManagement
 import akka.management.cluster.bootstrap.ClusterBootstrap
 import akka.stream.ActorMaterializer
 import org.apache.kafka.clients.producer.RecordMetadata
 import pureconfig._
-import org.apache.openwhisk.common.LoggingMarkers._
 import org.apache.openwhisk.common._
 import org.apache.openwhisk.core.WhiskConfig._
 import org.apache.openwhisk.core.connector._
 import org.apache.openwhisk.core.entity._
-import org.apache.openwhisk.core.entity.size._
+import org.apache.openwhisk.common.LoggingMarkers._
 import org.apache.openwhisk.core.loadBalancer.InvokerState.{Healthy, Offline, Unhealthy, Unresponsive}
 import org.apache.openwhisk.core.{ConfigKeys, WhiskConfig}
 import org.apache.openwhisk.spi.SpiLoader
 
 import scala.annotation.tailrec
-import scala.collection.concurrent.TrieMap
-import scala.concurrent.duration._
-import scala.concurrent.{ExecutionContext, Future, Promise}
-import scala.util.{Failure, Success}
+import scala.concurrent.Future
 
 /**
  * A loadbalancer that schedules workload based on a hashing-algorithm.
@@ -151,17 +144,13 @@ import scala.util.{Failure, Success}
 class ShardingContainerPoolBalancer(
   config: WhiskConfig,
   controllerInstance: ControllerInstanceId,
-  private val feedFactory: FeedFactory,
-  private val invokerPoolFactory: InvokerPoolFactory,
-  lbConfig: ShardingContainerPoolBalancerConfig =
-    loadConfigOrThrow[ShardingContainerPoolBalancerConfig](ConfigKeys.loadbalancer),
-  private val messagingProvider: MessagingProvider = SpiLoader.get[MessagingProvider])(
-  implicit val actorSystem: ActorSystem,
+  feedFactory: FeedFactory,
+  val invokerPoolFactory: InvokerPoolFactory,
+  implicit val messagingProvider: MessagingProvider = SpiLoader.get[MessagingProvider])(
+  implicit actorSystem: ActorSystem,
   logging: Logging,
   materializer: ActorMaterializer)
-    extends LoadBalancer {
-
-  private implicit val executionContext: ExecutionContext = actorSystem.dispatcher
+    extends CommonLoadBalancer(config, feedFactory, controllerInstance) {
 
   /** Build a cluster of all loadbalancers */
   private val cluster: Option[Cluster] = if (loadConfigOrThrow[ClusterConfig](ConfigKeys.cluster).useClusterBootstrap) {
@@ -174,45 +163,26 @@ class ShardingContainerPoolBalancer(
     None
   }
 
-  /** State related to invocations and throttling */
-  protected[loadBalancer] val activationSlots = TrieMap[ActivationId, ActivationEntry]()
-  protected[loadBalancer] val activationPromises =
-    TrieMap[ActivationId, Promise[Either[ActivationId, WhiskActivation]]]()
-  private val activationsPerNamespace = TrieMap[UUID, LongAdder]()
-  private val totalActivations = new LongAdder()
-  private val totalBlackBoxActivationMemory = new LongAdder()
-  private val totalManagedActivationMemory = new LongAdder()
-
-  /** State needed for scheduling. */
-  protected[loadBalancer] val schedulingState = ShardingContainerPoolBalancerState()(lbConfig)
-
-  actorSystem.scheduler.schedule(0.seconds, 10.seconds) {
-    MetricEmitter.emitHistogramMetric(LOADBALANCER_ACTIVATIONS_INFLIGHT(controllerInstance), totalActivations.longValue)
+  override protected def emitHistogramMetric() = {
+    super.emitHistogramMetric()
     MetricEmitter.emitHistogramMetric(
-      LOADBALANCER_MEMORY_INFLIGHT(controllerInstance, ""),
-      totalBlackBoxActivationMemory.longValue + totalManagedActivationMemory.longValue)
-    MetricEmitter.emitHistogramMetric(
-      LOADBALANCER_MEMORY_INFLIGHT(controllerInstance, "Blackbox"),
-      totalBlackBoxActivationMemory.longValue)
-    MetricEmitter.emitHistogramMetric(
-      LOADBALANCER_MEMORY_INFLIGHT(controllerInstance, "Managed"),
-      totalManagedActivationMemory.longValue)
-    MetricEmitter.emitHistogramMetric(INVOKER_TOTALMEM_BLACKBOX, schedulingState.blackboxInvokers.foldLeft(0L) {
-      (total, curr) =>
+      INVOKER_TOTALMEM_BLACKBOX,
+      schedulingState.blackboxInvokers.foldLeft(0L) { (total, curr) =>
         if (curr.status.isUsable) {
           curr.id.userMemory.toMB + total
         } else {
           total
         }
-    })
-    MetricEmitter.emitHistogramMetric(INVOKER_TOTALMEM_MANAGED, schedulingState.managedInvokers.foldLeft(0L) {
-      (total, curr) =>
+      })
+    MetricEmitter.emitHistogramMetric(
+      INVOKER_TOTALMEM_MANAGED,
+      schedulingState.managedInvokers.foldLeft(0L) { (total, curr) =>
         if (curr.status.isUsable) {
           curr.id.userMemory.toMB + total
         } else {
           total
         }
-    })
+      })
     MetricEmitter.emitHistogramMetric(
       HEALTHY_INVOKER_MANAGED,
       schedulingState.managedInvokers.count(_.status == Healthy))
@@ -238,6 +208,9 @@ class ShardingContainerPoolBalancer(
       OFFLINE_INVOKER_BLACKBOX,
       schedulingState.blackboxInvokers.count(_.status == Offline))
   }
+
+  /** State needed for scheduling. */
+  val schedulingState = ShardingContainerPoolBalancerState()(lbConfig)
 
   /**
    * Monitors invoker supervision and the cluster to update the state sequentially
@@ -283,9 +256,6 @@ class ShardingContainerPoolBalancer(
 
   /** Loadbalancer interface methods */
   override def invokerHealth(): Future[IndexedSeq[InvokerHealth]] = Future.successful(schedulingState.invokers)
-  override def activeActivationsFor(namespace: UUID): Future[Int] =
-    Future.successful(activationsPerNamespace.get(namespace).map(_.intValue()).getOrElse(0))
-  override def totalActiveActivations: Future[Int] = Future.successful(totalActivations.intValue())
   override def clusterSize: Int = schedulingState.clusterSize
 
   /** 1. Publish a message to the loadbalancer */
@@ -344,201 +314,19 @@ class ShardingContainerPoolBalancer(
       }
   }
 
-  /**
-   * 2. Update local state with the to be executed activation.
-   *
-   * All activations are tracked in the activationSlots map. Additionally, blocking invokes
-   * are tracked in the activation results map. When a result is received via activeack, it
-   * will cause the result to be forwarded to the caller waiting on the result, and cancel
-   * the DB poll which is also trying to do the same.
-   */
-  private def setupActivation(msg: ActivationMessage,
-                              action: ExecutableWhiskActionMetaData,
-                              instance: InvokerInstanceId): Future[Either[ActivationId, WhiskActivation]] = {
-
-    totalActivations.increment()
-    val isBlackboxInvocation = action.exec.pull
-    val totalActivationMemory =
-      if (isBlackboxInvocation) totalBlackBoxActivationMemory else totalManagedActivationMemory
-    totalActivationMemory.add(action.limits.memory.megabytes)
-
-    activationsPerNamespace.getOrElseUpdate(msg.user.namespace.uuid, new LongAdder()).increment()
-
-    // Timeout is a multiple of the configured maximum action duration. The minimum timeout is the configured standard
-    // value for action durations to avoid too tight timeouts.
-    // Timeouts in general are diluted by a configurable factor. In essence this factor controls how much slack you want
-    // to allow in your topics before you start reporting failed activations.
-    val timeout = (action.limits.timeout.duration.max(TimeLimit.STD_DURATION) * lbConfig.timeoutFactor) + 1.minute
-
-    val resultPromise = if (msg.blocking) {
-      activationPromises.getOrElseUpdate(msg.activationId, Promise[Either[ActivationId, WhiskActivation]]()).future
-    } else Future.successful(Left(msg.activationId))
-
-    // Install a timeout handler for the catastrophic case where an active ack is not received at all
-    // (because say an invoker is down completely, or the connection to the message bus is disrupted) or when
-    // the active ack is significantly delayed (possibly dues to long queues but the subject should not be penalized);
-    // in this case, if the activation handler is still registered, remove it and update the books.
-    activationSlots.getOrElseUpdate(
-      msg.activationId, {
-        val timeoutHandler = actorSystem.scheduler.scheduleOnce(timeout) {
-          processCompletion(msg.activationId, msg.transid, forced = true, isSystemError = false, invoker = instance)
-        }
-
-        // please note: timeoutHandler.cancel must be called on all non-timeout paths, e.g. Success
-        ActivationEntry(
-          msg.activationId,
-          msg.user.namespace.uuid,
-          instance,
-          action.limits.memory.megabytes.MB,
-          action.limits.concurrency.maxConcurrent,
-          action.fullyQualifiedName(true),
-          timeoutHandler,
-          isBlackboxInvocation)
-      })
-
-    resultPromise
-  }
-
-  private val messageProducer = messagingProvider.getProducer(config, Some(ActivationEntityLimit.MAX_ACTIVATION_LIMIT))
-
-  /** 3. Send the activation to the invoker */
-  private def sendActivationToInvoker(producer: MessageProducer,
-                                      msg: ActivationMessage,
-                                      invoker: InvokerInstanceId): Future[RecordMetadata] = {
-    implicit val transid: TransactionId = msg.transid
-
-    val topic = s"invoker${invoker.toInt}"
-
-    MetricEmitter.emitCounterMetric(LoggingMarkers.LOADBALANCER_ACTIVATION_START)
-    val start = transid.started(this, LoggingMarkers.CONTROLLER_KAFKA)
-
-    producer.send(topic, msg).andThen {
-      case Success(status) =>
-        transid.finished(
-          this,
-          start,
-          s"posted to ${status.topic()}[${status.partition()}][${status.offset()}]",
-          logLevel = InfoLevel)
-      case Failure(_) => transid.failed(this, start, s"error on posting to topic $topic")
-    }
-  }
-
-  /**
-   * Subscribes to active acks (completion messages from the invokers), and
-   * registers a handler for received active acks from invokers.
-   */
-  private val activationFeed: ActorRef =
-    feedFactory.createFeed(actorSystem, messagingProvider, processAcknowledgement)
-
-  /** 4. Get the active-ack message and parse it */
-  protected[loadBalancer] def processAcknowledgement(bytes: Array[Byte]): Future[Unit] = Future {
-    val raw = new String(bytes, StandardCharsets.UTF_8)
-    AcknowledegmentMessage.parse(raw) match {
-      case Success(m: CompletionMessage) =>
-        processCompletion(
-          m.activationId,
-          m.transid,
-          forced = false,
-          isSystemError = m.isSystemError,
-          invoker = m.invoker)
-        activationFeed ! MessageFeed.Processed
-
-      case Success(m: ResultMessage) =>
-        processResult(m.response, m.transid)
-        activationFeed ! MessageFeed.Processed
-
-      case Failure(t) =>
-        activationFeed ! MessageFeed.Processed
-        logging.error(this, s"failed processing message: $raw")
-
-      case _ =>
-        activationFeed ! MessageFeed.Processed
-        logging.error(this, s"Unexpected Acknowledgment message received by loadbalancer: $raw")
-    }
-  }
-
-  /** 5. Process the result ack and return it to the user */
-  private def processResult(response: Either[ActivationId, WhiskActivation], tid: TransactionId): Unit = {
-    val aid = response.fold(l => l, r => r.activationId)
-
-    // Resolve the promise to send the result back to the user.
-    // The activation will be removed from the activation slots later, when the completion message
-    // is received (because the slot in the invoker is not yet free for new activations).
-    activationPromises.remove(aid).foreach(_.trySuccess(response))
-    logging.info(this, s"received result ack for '$aid'")(tid)
-  }
-
-  /** Process the completion ack and update the state */
-  protected[loadBalancer] def processCompletion(aid: ActivationId,
-                                                tid: TransactionId,
-                                                forced: Boolean,
-                                                isSystemError: Boolean,
-                                                invoker: InvokerInstanceId): Unit = {
-
-    val invocationResult = if (forced) {
-      InvocationFinishedResult.Timeout
-    } else {
-      // If the response contains a system error, report that, otherwise report Success
-      // Left generally is considered a Success, since that could be a message not fitting into Kafka
-      if (isSystemError) {
-        InvocationFinishedResult.SystemError
-      } else {
-        InvocationFinishedResult.Success
-      }
-    }
-
-    activationSlots.remove(aid) match {
-      case Some(entry) =>
-        totalActivations.decrement()
-        val totalActivationMemory =
-          if (entry.isBlackbox) totalBlackBoxActivationMemory else totalManagedActivationMemory
-        totalActivationMemory.add(entry.memory.toMB * (-1))
-        activationsPerNamespace.get(entry.namespaceId).foreach(_.decrement())
-        schedulingState.invokerSlots
-          .lift(invoker.toInt)
-          .foreach(_.releaseConcurrent(entry.fullyQualifiedEntityName, entry.maxConcurrent, entry.memory.toMB.toInt))
-        if (!forced) {
-          entry.timeoutHandler.cancel()
-          // notice here that the activationPromises is not touched, because the expectation is that
-          // the active ack is received as expected, and processing that message removed the promise
-          // from the corresponding map
-        } else {
-          // the entry has timed out; if the active ack is still around, remove its entry also
-          // and complete the promise with a failure if necessary
-          activationPromises
-            .remove(aid)
-            .foreach(_.tryFailure(new Throwable("no completion or active ack received yet")))
-        }
-
-        logging.info(this, s"${if (!forced) "received" else "forced"} completion ack for '$aid'")(tid)
-        // Active acks that are received here are strictly from user actions - health actions are not part of
-        // the load balancer's activation map. Inform the invoker pool supervisor of the user action completion.
-        invokerPool ! InvocationFinishedMessage(invoker, invocationResult)
-      case None if tid == TransactionId.invokerHealth =>
-        // Health actions do not have an ActivationEntry as they are written on the message bus directly. Their result
-        // is important to pass to the invokerPool because they are used to determine if the invoker can be considered
-        // healthy again.
-        logging.info(this, s"received completion ack for health action on $invoker")(tid)
-        invokerPool ! InvocationFinishedMessage(invoker, invocationResult)
-      case None if !forced =>
-        // Received an active-ack that has already been taken out of the state because of a timeout (forced active-ack).
-        // The result is ignored because a timeout has already been reported to the invokerPool per the force.
-        logging.debug(this, s"received completion ack for '$aid' which has no entry")(tid)
-      case None =>
-        // The entry has already been removed by an active ack. This part of the code is reached by the timeout and can
-        // happen if active-ack and timeout happen roughly at the same time (the timeout was triggered before the active
-        // ack canceled the timer). As the active ack is already processed we don't have to do anything here.
-        logging.debug(this, s"forced completion ack for '$aid' which has no entry")(tid)
-    }
-  }
-
-  private val invokerPool =
+  override val invokerPool =
     invokerPoolFactory.createInvokerPool(
       actorSystem,
       messagingProvider,
       messageProducer,
       sendActivationToInvoker,
       Some(monitor))
+
+  override protected def releaseInvoker(invoker: InvokerInstanceId, entry: ActivationEntry) = {
+    schedulingState.invokerSlots
+      .lift(invoker.toInt)
+      .foreach(_.releaseConcurrent(entry.fullyQualifiedEntityName, entry.maxConcurrent, entry.memory.toMB.toInt))
+  }
 }
 
 object ShardingContainerPoolBalancer extends LoadBalancerProvider {
@@ -547,24 +335,6 @@ object ShardingContainerPoolBalancer extends LoadBalancerProvider {
     implicit actorSystem: ActorSystem,
     logging: Logging,
     materializer: ActorMaterializer): LoadBalancer = {
-
-    val activeAckTopic = s"completed${instance.asString}"
-    val maxActiveAcksPerPoll = 128
-    val activeAckPollDuration = 1.second
-
-    val feedFactory = new FeedFactory {
-      def createFeed(f: ActorRefFactory, provider: MessagingProvider, acker: Array[Byte] => Future[Unit]) = {
-        f.actorOf(Props {
-          new MessageFeed(
-            "activeack",
-            logging,
-            provider.getConsumer(whiskConfig, activeAckTopic, activeAckTopic, maxPeek = maxActiveAcksPerPoll),
-            maxActiveAcksPerPoll,
-            activeAckPollDuration,
-            acker)
-        })
-      }
-    }
 
     val invokerPoolFactory = new InvokerPoolFactory {
       override def createInvokerPool(
@@ -585,7 +355,11 @@ object ShardingContainerPoolBalancer extends LoadBalancerProvider {
       }
 
     }
-    new ShardingContainerPoolBalancer(whiskConfig, instance, feedFactory, invokerPoolFactory)
+    new ShardingContainerPoolBalancer(
+      whiskConfig,
+      instance,
+      createFeedFactory(whiskConfig, instance),
+      invokerPoolFactory)
   }
 
   def requiredProperties: Map[String, String] = kafkaHosts
@@ -613,7 +387,6 @@ object ShardingContainerPoolBalancer extends LoadBalancerProvider {
    *
    * @param maxConcurrent concurrency limit supported by this action
    * @param invokers a list of available invokers to search in, including their state
-   * @param concurrentSlots optional map of invoker -> semaphore to track concurrency slots for this action
    * @param dispatched semaphores for each invoker to give the slots away from
    * @param slots Number of slots, that need to be acquired (e.g. memory in MB)
    * @param index the index to start from (initially should be the "homeInvoker"
