+++ tests/src/test/scala/org/apache/openwhisk/core/loadBalancer/test/ShardingContainerPoolBalancerTests.scala
@@ -476,13 +476,7 @@ class ShardingContainerPoolBalancerTests
         TestProbe().testActor
     }
     val balancer =
-      new ShardingContainerPoolBalancer(
-        config,
-        ControllerInstanceId("0"),
-        feedProbe,
-        invokerPoolProbe,
-        lbConfig(0.0),
-        mockMessaging)
+      new ShardingContainerPoolBalancer(config, ControllerInstanceId("0"), feedProbe, invokerPoolProbe, mockMessaging)
 
     val invokers = IndexedSeq.tabulate(numInvokers) { i =>
       new InvokerHealth(InvokerInstanceId(i, userMemory = invokerMem), Healthy)
